// src/types/api-requests.ts
interface ResponseBody {
  success: boolean;
  message?: string;
  token?: string;
  role?: string;
  [key: string]: unknown;
}
// ...
interface ColorsRequestBody {
  page?: number;
  resultsOnPage?: number;
  code?: string;
  name_uk?: string;
  name_en?: string;
  sortField?: string;
  sortDesc?: boolean;
}
interface ColorsResponseBody extends ResponseBody {
  colors?: Color[];
  count?: number;
}
interface AddColorRequestBody {
  color: Color;
}
interface UpdateColorRequestBody {
  id: string;
  newName_uk: string;
  newName_en: string;
  newHex: string;
}
interface RemoveColorRequestBody {
  id: string;
}